add intel sg-size to 16

+++ src/Types/GridTerm.h
@@ -23,7 +23,7 @@ namespace ARBD {
  *          (see BaseGrid<T>::Config::boundary_as_int) so the struct stays POD
  *          across the CUDA/SYCL/Metal backends.
  */
-enum class GridBoundaryCondition : int {
+enum class GridBoundaryCondition : uint8_t {
 	Dirichlet = 0, ///< Fixed value at boundary (legacy zero-pads out-of-bounds taps)
 	Neumann = 1,   ///< Fixed derivative at boundary
 	Periodic = 2   ///< Periodic boundary conditions
@@ -32,9 +32,8 @@ enum class GridBoundaryCondition : int {
 /// Storage backend a grid is held in. grid_ids are unified across both.
 enum class GridFormat : uint8_t { Dense = 0, Sparse = 1 };
 
-
 /// Interpolation order used when sampling a grid (legacy's `scheme` flag).
-enum class InterpolationOrder : int { Linear = 1, Cubic = 3 };
+enum class InterpolationOrder : uint8_t { Linear = 1, Cubic = 3 };
 
 /**
  * @brief One (entity type -> grid) reference: which grid, and how it is weighted
@@ -42,9 +41,7 @@ enum class InterpolationOrder : int { Linear = 1, Cubic = 3 };
  *          whitespace-separated list with one scale and one boundary condition
  *          per entry - so per-type grid data is stored as an offset+count range
  *          into a flat array of these terms (see ParticleTypeView::pmf_grid_terms
- *          and RigidBodyTypeView's grid ranges). Deliberately a 16-byte POD: a
- *          grid-force kernel reads one of these per grid in its inner loop, and
- *          a single load beats four parallel-array loads.
+ *          and RigidBodyTypeView's grid ranges).
  *
  *          `grid_id` indexes GridManager's per-resource grid view array rather
  *          than inlining a view here, for two reasons. It keeps the term
@@ -56,9 +53,9 @@ enum class InterpolationOrder : int { Linear = 1, Cubic = 3 };
  *          uploaded.
  */
 struct GridTerm {
-	int grid_id = -1;			 ///< Index into GridManager's grid views; <0 = unused slot
-	float scale = 1.0f;			 ///< legacy: gridFileScale
-	float scale_slope = 0.0f;	 ///< legacy: gridFileScaleSlope
+	int grid_id = -1;		  ///< Index into GridManager's grid views; <0 = unused slot
+	float scale = 1.0f;		  ///< legacy: gridFileScale
+	float scale_slope = 0.0f; ///< legacy: gridFileScaleSlope
 	/// legacy: gridFileBoundaryConditions; <0 = use the grid's own.
 	/// Recorded per (type, grid) because that's how the config keys it, but not
 	/// yet observable: BaseGridDevice.h's samplers take a boundary_condition
